- feat: update json payload

+++ src/main/java/com/paisabazaar/kafka_producer/controller/ProducerController.java
@@ -8,6 +8,7 @@ import com.paisabazaar.kafka_producer.service.ResponseFormatter;
 import com.paisabazaar.kafka_producer.utils.ResourceNotFoundException;
 import com.paisabazaar.kafka_producer.utils.ResponseCode;
 import org.json.JSONObject;
+import org.springframework.beans.factory.annotation.Required;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.handler.annotation.Header;
@@ -38,17 +39,23 @@ public class ProducerController {
     }
 
     @GetMapping(value = "/producer", produces = "application/json")
-    public ResponseEntity<HashMap<String, Object>> getProducers() {
+    public ResponseEntity<?> getProducers() {
         HashMap<String, Object> producersMap = new HashMap<>();
         producerRepository.findAll().forEach(producer -> {
             producersMap.put(producer.getId(), producer);
         });
         // ## Need to implement custom exception handler and custom response handler
-        return new ResponseEntity<>(producersMap, HttpStatus.OK);
+        JSONObject data = new JSONObject();
+        data.put("producer", producersMap);
+        JSONObject response = responseFormatter.buildResponse(
+                "success", HttpStatus.OK.value(),
+                data.getJSONObject("producer"),
+                "Fetched successfully");
+        return new ResponseEntity<>(response.toMap(), HttpStatus.OK);
     }
 
     @PostMapping(value = "/producer", produces = "application/json")
-    public ResponseEntity<Producer> createProducer(@RequestBody Producer producer) {
+    public ResponseEntity<?> createProducer(@RequestBody Producer producer) {
         // Set uuid to id
         producer.setId(UUID.randomUUID().toString());
         // Set createdAt and updatedAt timestamps
@@ -58,29 +65,63 @@ public class ProducerController {
         // Save producer in redis
         producerRepository.save(producer);
         // ## Need to implement custom exception handler and custom response handler
-        return new ResponseEntity<>(producer, HttpStatus.CREATED);
+        JSONObject data = new JSONObject();
+        data.put("producer", producer);
+        JSONObject response = responseFormatter.buildResponse(
+                "success", HttpStatus.OK.value(),
+                data,
+                "Producer created");
+        return new ResponseEntity<>(response.toMap(), HttpStatus.CREATED);
     }
 
     @PutMapping(value = "/producer/{id}", produces = "application/json")
-    public ResponseEntity<Producer> updateProducer(@PathVariable String id, @RequestBody Producer payload) {
+    public ResponseEntity<?> updateProducer(@PathVariable String id, @RequestBody Producer payload) {
         // Update producer
-        Producer p = producerRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Producer", "id", id));
+        if (!producerRepository.existsById(id)) {
+            JSONObject response = responseFormatter.buildErrorResponse(
+                    "error",
+                    ResponseCode.PRODUCER_NOT_RETRIEVED.getCode(),
+                    new JSONObject().put("message", ResponseCode.PRODUCER_NOT_RETRIEVED.getMessage())
+            );
+            return new ResponseEntity<>(response.toMap(), HttpStatus.NOT_FOUND);
+        }
+        Producer p = producerRepository.findById(id).get();
         applicationUtilsService.copyNonNullProperties(payload, p);
         p.setUpdatedAt(new Date());
         producerRepository.save(p);
         // ## Need to implement custom exception handler and custom response handler
-        return new ResponseEntity<>(p, HttpStatus.CREATED);
+        JSONObject data = new JSONObject();
+        data.put("producer", p);
+        JSONObject response = responseFormatter.buildResponse(
+                "success", HttpStatus.OK.value(),
+                data,
+                "Producer updated");
+        return new ResponseEntity<>(response.toMap(), HttpStatus.CREATED);
     }
 
     @DeleteMapping(value = "/producer/{id}", produces = "application/json")
-    public @ResponseBody String deleteProducer(@PathVariable String id) {
+    public ResponseEntity<?> deleteProducer(@PathVariable String id) {
         // ## Need to implement custom exception handler and custom response handler
-        producerRepository.deleteById(id);
-        return "{'message': 'Deleted successfully.'}";
+        if (producerRepository.existsById(id)) {
+            producerRepository.deleteById(id);
+            JSONObject response = responseFormatter.buildResponse(
+                    "success",
+                    HttpStatus.OK.value(),
+                    null,
+                    "Producer Deleted");
+            return new ResponseEntity<>(response.toMap(), HttpStatus.OK);
+        } else {
+            JSONObject response = responseFormatter.buildErrorResponse(
+                    "error",
+                    ResponseCode.PRODUCER_NOT_RETRIEVED.getCode(),
+                    new JSONObject().put("message", ResponseCode.PRODUCER_NOT_RETRIEVED.getMessage())
+            );
+            return new ResponseEntity<>(response.toMap(), HttpStatus.NOT_FOUND);
+        }
     }
 
     @PostMapping(value = "/produce_messages", produces = "application/json")
-    public ResponseEntity<?> produceMessage(@RequestHeader(value = "x-producer-id") String id, @RequestBody String message) throws ExecutionException, InterruptedException {
+    public ResponseEntity<?> produceMessages(@RequestHeader(value = "x-producer-id") String id, @RequestBody String message) throws ExecutionException, InterruptedException {
         // ## Get Topic from producer id
         String topic = producerRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Producer", "id", id)).getTopic();
         // ## Produce to kafka
